pythonProject-Django/requirements.txt
asgiref==3.4.1
distlib==0.3.4
Django==3.2.12
djangorestframework==3.13.1
filelock==3.4.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
platformdirs==2.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
virtualenv==20.13.1
zipp==3.6.0
